refactor: update auth imports and restructure risk tools
- Replaced instances of the auth import from "@/utils/auth" to "@/utils/auth-client" in remove-member.ts and route.ts for improved modularity and consistency.
- Introduced a new risks-tool.ts file to encapsulate risk-related functionalities, replacing the deprecated risks.ts file.
- Ensured all related functions maintain type safety and adhere to the modular structure of the codebase.
