+++ apps/app/src/data/tools/index.ts
@@ -1,6 +1,6 @@
 import { getOrganizationTools } from "./organization";
 import { getPolicyTools } from "./policies";
-import { getRiskTools } from "./risks";
+import { getRiskTools } from "./risks-tool";
 import { getUserTools } from "./user";
 
 export const tools = {
